refactor(router): move empty outlet component
Due to future circular references between RouterOutlet and EmptyOutletComponent, the EmptyOutletComponent has been moved to the RouterOutlet file.
