+++ packages/router/src/private_export.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-export {ɵEmptyOutletComponent} from './components/empty_outlet';
+export {EmptyOutletComponent as ɵEmptyOutletComponent} from './directives/router_outlet';
 export {RestoredState as ɵRestoredState} from './navigation_transition';
 export {loadChildren as ɵloadChildren} from './router_config_loader';
 export {ROUTER_PROVIDERS as ɵROUTER_PROVIDERS} from './router_module';
